Refactor permission sets for seller and admin
