+++ lib/spree/permission_sets/admin.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Spree
+  module PermissionSets
+    class Admin < PermissionSets::Base
+      def activate!
+        cannot :visit, :seller_dashboard unless user&.has_spree_role?('seller')
+        cannot :visit, :seller_prices unless user&.has_spree_role?('seller')
+        cannot :visit, :paypal_callbacks unless user&.has_spree_role?('seller')
+      end
+    end
+  end
+end
